+++ src/main/java/ru/vddmit/task_db_01/models/OrderItem.java
@@ -16,7 +16,7 @@ public class OrderItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "order_item_id")
-    private int orderItemId;
+    private long orderItemId;
     @ManyToOne
     @JoinColumn(name = "order_id", nullable = false)
     private Order order;
